Reorganizate directories ...
