+++ src/controllers/register.controller.ts
@@ -2,7 +2,7 @@
 
 import dotenv from 'dotenv';
 import {validateNewUser} from './validations/controller.validations';
-import { UserRepository } from '../repository/user.repository';
+import { UserRepository } from '../repository/implementations/user.implementation';
 import * as AuthUtil from '../utils/AuthUtil.utils';
 import { AuthMessages } from '../utils/AuthMessage.util';
